Nest CSharpCodeFixTest`2 within CSharpCodeFixVerifier`2
